Move test for bounty_url_for to different module and adjust other tests to changes introduced on master
